Add ui_rect_to_retina() and ui_rect_from_retina() for now
However, if you use any of the built-in drawing tools in this rect,
it will scale as if retina is still active, so it will be in the
wrong place.

What we might need to do is to store the retina scaling factor as
a property of the viewport, so we can set it to 1 here.  Built-in
controls would draw at the incorrect (non-retina) scale, but in
the correct place.